Some fixes

+++ app/src/main/java/com/cmput301fa16t5/legendary_telegram/UserProfileController.java
@@ -121,7 +121,7 @@ public class UserProfileController {
     private UserProfileEnum checkUserInput(String name, String email, String phone) {
 
         try{
-            Integer.parseInt(phone.replaceAll("-", "").replaceAll(" ", ""));
+            Double.parseDouble(phone.replaceAll("-", "").replaceAll(" ", ""));
         }
 
         catch (NumberFormatException e) {
